Split flaker batch VRT extension
